Docs(templating): Issue and PR templates (#22)
* docs(typo fix): fix a typo

* docs(issue template): add issue templates

This change will add issue templates for bug reports and feature
requests. They include subheadings to include along with a checklist for
developers to follow to ensure developers working on the issue have
detailed steps to reproduce the issue or understand the desired feature.

* docs(PR template): adds a Pull Request template

This template enforces contributors to detail their changes and why they
are needed, furthermore, the checklist classifies the change, documents
its linked issus(s), ensures the change is documented, and that tests
have been written for the change.

* docs: document pr and issue structure

* docs(format): adjust grammar & template structure

this will fix grammar issues and the general template structure of the
bug reports, feature requests, and pull request templates.

* docs(wording): adjust template wording

This change helps make the template more readable and have text flow
nicer.