refactor(file structure): standardize volume-cleaner's structure (#48)
* refactor: organize files into internal folder

Instead of using one package, files are now seperated into there functional folders, the kubernetes package in internal is used for interaction with the kubernetes server like retrieving resources (sts, pvcs, etc.)

The Structures folder is for data structures, in this case, the set.go file has been moved in since a set is a data structure

Futhermore, an additional function was added (A get set function) since the complier was complaining when the project was seperated into packages, it stated unattachedPVCs.list was a private field. So a getter was created in the set.go function (This getter function doesn't follow good practice which is why it has been tagged with a HACK comment)

* refactor: rename a getter from getList to getSet

The underlying structure is a list, but we don't want to tell the user
that, they are getting a "set".

* refactor(configs): add a configs directory

* refactor(images): rename public folder to assets

* refactor(file name): change to snake_case

currently file and folder names have a mixture of snake_case and
kebab-case, the purpose of this change is to ensure all file and folder
names follow snake case as that is the common practice

* refactor(submakes): create a scripts folder

In order to prevent clutter in the main makefile, the scripts folder for
builds will house submake files for the project

* docs(file structure): document file structure

The file structure has been documented in the README.md for future
contributors, it follows the golang-standard project layout, file and
directories are written in snake_case.

* refactor(rename): change variable name

kuber is a confusing name especially with a local variable kube, this
change clarifies we have an internal package called kubernetes as well.

---------

Co-authored-by: Justin Zhang <justinjing.zhang@statcan.gc.ca>