feat(control loop): add the control loop of the watcher (#54)
* feat(watcher): add a main control loop

This change will add a main control loop that watches for statefulset
events and triggers on sts creation or deletion

* fix(packages): import missing packages

The binary for the volume-cleaner was failing to compile due to unknown operators
(OR operator) and packages such as Appsv1 and kubernetes package not
being imported. This commit fixes that.

* feat(sts pvcs): get the pvcs of a sts

This change will create a function in the kubernetes/retriever package
that given a sts, it will return its associated persistent volume
claims, this is done using owner references, there are alternative
methods like using the name structure of PVCs, but that makes it less
robust, thus the owner reference method was use. Unit testing has also
been done to test whether the function acquires the right owned pvcs and
ignores the ones that are not "owned"

* feat(sts pvc): add get pvcs of sts in control loop

This adds getting the pvcs associated with an sts into the main
control loop of the watcher

* fix(remove): remove or comment variables for tests

* fix(return pointer): return a slice

Due to escape analysis, it is bad to return a slice pointer, this has
been fixed now

---------

Co-authored-by: Justin Zhang <justinjing.zhang@statcan.gc.ca>