fix: remove line that is flooding logs
Related to #131, it seems like this line is continuously flooding the controller logs. We've never had
this issue before, which leads me to think that it's not even a bug. Maybe it's just expected behaviour.
Regardless, I'm removing it for now to stop the spam, and we'll see if we encounter further issues.