refactor: improve code hygiene (#128)
* fix: remove previous placeholder comments

* fix: remove redundant variable declarations

* refactor: make kube client init function shared

move the func to internal package to prevent code duplication, whereas
previously it was declared in both main.go files

* refactor: improve code flow of finder.go

* refactor: improve logs for watcher.go

* refactor: edit comment

* refactor: move structs from types.go to config.go

Because the contents of both files are just config-like structs, I didn't feel like
it was necessary to have 2 files for that. I also didn't like the generic name of
types.go.

* refactor: rename and move unit testing obj

being in a directory called tests was confusing (could be mistaken for unit tests folder), and
being called client.go conflicts with the kube client package name (even though there are no issues,
it's still unecessary confusion)

* fix: fix integration test

* refactor: improve logs

* feat: add makefile command to manually start jobs

* fix: fix log STS->NS

* fix: remove line that is flooding logs

Related to #131, it seems like this line is continuously flooding the controller logs. We've never had
this issue before, which leads me to think that it's not even a bug. Maybe it's just expected behaviour.
Regardless, I'm removing it for now to stop the spam, and we'll see if we encounter further issues.